extension: Update status icon to service icon
Service state is updated to status icon area.
When service state is "associating" or "configuring" the status
icon update to the particular technology icon, but for rest of the
cases it shows the default service icon. If this icon is wifi or
cellular, the signal strength is reflected as well.

Some minor bugfixes go in here as well.